requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
tweepy==4.14.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
apscheduler==3.10.4
python-multipart==0.0.9
pydantic==2.6.1
